+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/FluidPump.java
@@ -4,8 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 import org.bukkit.Material;
+import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
+import org.bukkit.block.data.BlockData;
+import org.bukkit.block.data.Levelled;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.ItemStack;
@@ -27,10 +30,17 @@ import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces.InventoryBlock;
 import me.mrCookieSlime.Slimefun.Objects.handlers.BlockTicker;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
-import me.mrCookieSlime.Slimefun.api.energy.ChargableBlock;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenuPreset;
 
+/**
+ * This machine collects liquids from the {@link World} and puts them
+ * into buckets provided to the machine by using energy.
+ *
+ * @author TheBusyBiscuit
+ * @author Linox
+ *
+ */
 public class FluidPump extends SimpleSlimefunItem<BlockTicker> implements InventoryBlock, EnergyNetComponent {
 
     private static final int ENERGY_CONSUMPTION = 32;
@@ -59,14 +69,14 @@ public class FluidPump extends SimpleSlimefunItem<BlockTicker> implements Invent
 
     private void constructMenu(BlockMenuPreset preset) {
         for (int i : border) {
-            preset.addItem(i, new CustomItem(new ItemStack(Material.GRAY_STAINED_GLASS_PANE), " "), ChestMenuUtils.getEmptyClickHandler());
+            preset.addItem(i, new CustomItem(Material.GRAY_STAINED_GLASS_PANE, " "), ChestMenuUtils.getEmptyClickHandler());
         }
 
         for (int i : inputBorder) {
-            preset.addItem(i, new CustomItem(new ItemStack(Material.CYAN_STAINED_GLASS_PANE), " "), ChestMenuUtils.getEmptyClickHandler());
+            preset.addItem(i, new CustomItem(Material.CYAN_STAINED_GLASS_PANE, " "), ChestMenuUtils.getEmptyClickHandler());
         }
         for (int i : outputBorder) {
-            preset.addItem(i, new CustomItem(new ItemStack(Material.ORANGE_STAINED_GLASS_PANE), " "), ChestMenuUtils.getEmptyClickHandler());
+            preset.addItem(i, new CustomItem(Material.ORANGE_STAINED_GLASS_PANE, " "), ChestMenuUtils.getEmptyClickHandler());
         }
 
         for (int i : getOutputSlots()) {
@@ -109,16 +119,16 @@ public class FluidPump extends SimpleSlimefunItem<BlockTicker> implements Invent
         Block fluid = b.getRelative(BlockFace.DOWN);
         Optional<ItemStack> bucket = getFilledBucket(fluid);
 
-        if (bucket.isPresent() && ChargableBlock.getCharge(b) >= ENERGY_CONSUMPTION) {
+        if (bucket.isPresent() && getCharge(b.getLocation()) >= ENERGY_CONSUMPTION) {
             BlockMenu menu = BlockStorage.getInventory(b);
 
             for (int slot : getInputSlots()) {
-                if (SlimefunUtils.isItemSimilar(menu.getItemInSlot(slot), new ItemStack(Material.BUCKET), true)) {
+                if (SlimefunUtils.isItemSimilar(menu.getItemInSlot(slot), new ItemStack(Material.BUCKET), true, false)) {
                     if (!menu.fits(bucket.get(), getOutputSlots())) {
                         return;
                     }
 
-                    ChargableBlock.addCharge(b, -ENERGY_CONSUMPTION);
+                    removeCharge(b.getLocation(), ENERGY_CONSUMPTION);
                     menu.consumeItem(slot);
                     menu.pushItem(bucket.get().clone(), getOutputSlots());
                     consumeFluid(fluid);
@@ -130,24 +140,47 @@ public class FluidPump extends SimpleSlimefunItem<BlockTicker> implements Invent
     }
 
     private void consumeFluid(Block fluid) {
-        if (fluid.getType() == Material.WATER) {
-            fluid.setType(Material.AIR);
-            return;
+        if (fluid.isLiquid()) {
+            if (fluid.getType() == Material.WATER || fluid.getType() == Material.BUBBLE_COLUMN) {
+                // With water we can be sure to find an infinite source whenever we go
+                // further than 2 blocks, so we can just remove the water here and save
+                // outselves all of that computing...
+                fluid.setType(Material.AIR);
+            }
+            else {
+                List<Block> list = Vein.find(fluid, RANGE, block -> isLiquid(block) && block.getType() == fluid.getType());
+                list.get(list.size() - 1).setType(Material.AIR);
+            }
         }
-
-        List<Block> list = Vein.find(fluid, RANGE, block -> block.isLiquid() && block.getType() == fluid.getType());
-        list.get(list.size() - 1).setType(Material.AIR);
     }
 
     private Optional<ItemStack> getFilledBucket(Block fluid) {
         if (fluid.getType() == Material.LAVA) {
             return Optional.of(new ItemStack(Material.LAVA_BUCKET));
         }
-        else if (fluid.getType() == Material.WATER) {
+        else if (fluid.getType() == Material.WATER || fluid.getType() == Material.BUBBLE_COLUMN) {
             return Optional.of(new ItemStack(Material.WATER_BUCKET));
         }
+        else {
+            return Optional.empty();
+        }
+    }
 
-        return Optional.empty();
+    private boolean isLiquid(Block block) {
+        if (block.isLiquid()) {
+            BlockData data = block.getBlockData();
+
+            if (data instanceof Levelled) {
+                // Check if this is a full block.
+                return ((Levelled) data).getLevel() == 0;
+            }
+            else {
+                return false;
+            }
+        }
+        else {
+            return false;
+        }
     }
 
     @Override